% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_dataframe.R
\name{balance_cut}
\alias{balance_cut}
\title{cut the dataframe into training and testing parts}
\usage{
balance_cut(df, prop, targ)
}
\arguments{
\item{df}{the dataframe}

\item{prop}{the proportion of the tesing datas}

\item{targ}{the label variable}
}
\description{
This function cut the dataframe into training and testing parts
with balancing the previous ratio of the positive and negtive samples
}
\examples{
test.position = balance_cut(p2p,0.25,"def")
traindf<-df[-test.position,]
testdf<-df[test.position,]
}

