% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_dataframe.R
\name{cv_balance}
\alias{cv_balance}
\title{get a cross validation data pairs}
\usage{
cv_balance(df, folds, targ)
}
\arguments{
\item{df}{the dataframe}

\item{folds}{the numbers of pairs}

\item{targ}{the label variable}
}
\description{
This function will  get a cross validation data pairs
}
\examples{
cv.list<-cv_balance(p2p,4,"def")
traindf<-df[-cv.list[[1]],]
testdf<-df[cv.list[[1]],]
}

