% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessPlot.R
\name{get_auc}
\alias{get_auc}
\title{computing the auc}
\usage{
get_auc(prob, target)
}
\arguments{
\item{prob}{predicted target probilities}

\item{target}{the true class of the test data}
}
\description{
computing the auc
}
\examples{
get_auc(prob,target)
}

