% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessPlot.R
\name{ks}
\alias{ks}
\title{computing the ks and plot}
\usage{
ks(prob, target, plots = FALSE, rat = NULL)
}
\arguments{
\item{prob}{predicted target probilities}

\item{target}{the true class of the test data}
}
\description{
computing the ks value of a predicted model and can plot graph
}
\examples{
ks(prob,target)
ks(prob,target,plot = T)
}

