% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessPlot.R
\name{liftChart}
\alias{liftChart}
\title{plot the f1 score curve}
\usage{
liftChart(prob, target, nums = 10)
}
\arguments{
\item{prob}{predicted target probilities}

\item{target}{the true class of the test data}
}
\description{
plot the f1 score curve
}
\examples{
f1plot(prob,target)
}

