% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_wrapper.R
\name{para_logic}
\alias{para_logic}
\title{the wrappered logic regression model}
\usage{
para_logic(df, useful, targ, scorer, nfolds = 4)
}
\arguments{
\item{df}{the dataframe}

\item{useful}{used features}

\item{targ}{the label variable}

\item{nfolds}{the cross validation numbers}
}
\description{
the wrappered logic regression model
}
\examples{

para_logic(df =p2p.model,useful,targ="def",nfolds=4)
}

