% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_wrapper.R
\name{pmult_logic}
\alias{pmult_logic}
\title{the major component of para_logic}
\usage{
pmult_logic(cl, df, useful, targ, scorer, cv.list)
}
\arguments{
\item{df}{the dataframe}

\item{useful}{used features}

\item{targ}{the label variable}
}
\description{
the major component of para_logic
}
\examples{

pmult_logic(df =p2p.model,useful,targ="def")
}

