% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessPlot.R
\name{rocplot}
\alias{rocplot}
\title{plot the roc curve}
\usage{
rocplot(prob, target, smooth = T)
}
\arguments{
\item{prob}{predicted target probilities}

\item{target}{the true class of the test data}
}
\description{
plot the roc curve
}
\examples{
rocplot(prob,target)
}

