% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessPlot.R
\name{S_Asse}
\alias{S_Asse}
\title{get TP,FP,TN,FN}
\usage{
S_Asse(prob, target)
}
\arguments{
\item{prob}{predicted target probilities}

\item{target}{the true class of the test data}
}
\description{
get TP,FP,TN,FN
}
\examples{
S_Asse(prob,target)
}

